+++ src/Components/HomeContainer/HomeContainer.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react'
 import { ViewContext } from '../../Context/ViewContext'
 import { VideoProvider } from '../../Context/VideoContext'
 import ContentDelivery from '../ContentDelivery/ContentDelivery'
-import CountdownTimer from '../Timer/Timer'
+import CountdownTimer from '../CountdownTimer/CountdownTimer'
 import ContentSelection from '../ContentSelection/ContentSelection'
 import style from './HomeContainer.module.scss'
